requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.0.0
